requirements.txt
flask==3.0.3
flask-socketio==5.3.6
supabase==2.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1
psutil==6.0.0
flask-cors==5.0.0
eventlet==0.36.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3